+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - A complete example app based on the [order-process workflow from Zeebe's "Getting Started Tutorial"][order_process].
+### Changed
+- `Sync + UnwindSafe` is now enforced on job handlers and all work is done on a threadpool with `futures-cpupool`.
 
 ## [0.18.0] - 2019-06-10
 Initial release of zeebest 🥏